requirements.txt
Django==1.11.6
Markdown==2.6.9
argparse==1.2.1
certifi==2017.7.27.1
chardet==3.0.4
django-filter==1.0.4
djangorestframework==3.7.0
idna==2.6
pytz==2017.2
requests==2.18.4
urllib3==1.22
wsgiref==0.1.2
